// types.ts
export type ThemeName = 'purple' | 'green' | 'lightBlue' | 'sunset' | 'islamic';

export interface Theme {
  bg: string;
  card: string;
  text: string;
  title: string;
  primaryButton: string;
  secondaryButton: string;
  input: string;
  activeTab: string;
  inactiveTab: string;
  modalBg: string;
  modalContent: string;
}